+++ testdata/codegen_matrix/fixtures/feature_stress/src/graphql/resolvers/decorator_query.ts
@@ -1,0 +1,21 @@
+import { gqlQuery } from "@snowtop/ent/graphql";
+
+export class DecoratedRuntimeResolver {
+  async runtimeReady(): Promise<boolean> {
+    return true;
+  }
+}
+
+gqlQuery({
+  class: "DecoratedRuntimeResolver",
+  name: "decoratedRuntimeReady",
+  type: Boolean,
+  async: true,
+})(
+  DecoratedRuntimeResolver.prototype,
+  "runtimeReady",
+  Object.getOwnPropertyDescriptor(
+    DecoratedRuntimeResolver.prototype,
+    "runtimeReady",
+  ),
+);